refactor(execucao): renomear salvar execucao para nomes semanticos
