+++ web/chamados/templates/execucao/chamado_execucao.html
@@ -54,7 +54,7 @@
   {{ block.super }}
   <script src="{% static 'execucao/js/execution_state.js' %}" defer></script>
   <script src="{% static 'execucao/js/execucao_detalhe.js' %}" defer></script>
-  <script src="{% static 'execucao/js/execucao_salvar.js' %}" defer></script>
+  <script src="{% static 'execucao/js/chamado_salvar_execucao_ajax.js' %}" defer></script>
   <script src="{% static 'execucao/js/execucao_item_configurado.js' %}" defer></script>
   <script src="{% static 'execucao/js/execucao_finalizar.js' %}" defer></script>
 {% endblock %}
