+++ web/chamados/templates/execucao/components/_salvar_execucao_bar.html
@@ -18,7 +18,7 @@
           id="btn-salvar-execucao"
           type="button"
           class="btn btn-primary"
-          data-url="{% url 'execucao:chamado_salvar_execucao' chamado.id %}"
+          data-url="{% url 'execucao:chamado_salvar_execucao_ajax' chamado.id %}"
         >
           Salvar execução
         </button>
